refactor: move init rebuild logic
